+++ chia/simulator/simulator_constants.py
@@ -1,5 +1,5 @@
 if __name__ == "__main__":
-    from chia.util.block_tools import BlockTools, test_constants
+    from tests.block_tools import BlockTools, test_constants
     from chia.util.default_root import DEFAULT_ROOT_PATH
 
     # TODO: mariano: fix this with new consensus
