+++ chia/simulator/start_simulator.py
@@ -6,7 +6,7 @@ from chia.full_node.full_node import FullNode
 from chia.rpc.full_node_rpc_api import FullNodeRpcApi
 from chia.server.outbound_message import NodeType
 from chia.server.start_service import run_service
-from chia.util.block_tools import BlockTools, test_constants
+from tests.block_tools import BlockTools, test_constants
 from chia.util.config import load_config_cli
 from chia.util.default_root import DEFAULT_ROOT_PATH
 from chia.util.path import mkdir, path_from_root
