+++ tests/core/daemon/test_daemon.py
@@ -6,7 +6,7 @@ import pytest
 from chia.server.outbound_message import NodeType
 from chia.server.server import ssl_context_for_server
 from chia.types.peer_info import PeerInfo
-from chia.util.block_tools import BlockTools
+from tests.block_tools import BlockTools
 from chia.util.ints import uint16
 from chia.util.ws_message import create_payload
 from tests.core.node_height import node_height_at_least
