+++ tests/core/full_node/test_performance.py
@@ -17,7 +17,7 @@ from chia.types.condition_with_args import ConditionWithArgs
 from chia.types.unfinished_block import UnfinishedBlock
 from chia.util.clvm import int_to_bytes
 from chia.util.ints import uint64
-from chia.util.wallet_tools import WalletTool
+from tests.wallet_tools import WalletTool
 from tests.core.fixtures import empty_blockchain  # noqa: F401
 
 from tests.connection_utils import add_dummy_connection, connect_and_get_peer
