+++ tests/core/test_farmer_harvester_rpc.py
@@ -14,7 +14,7 @@ from chia.rpc.harvester_rpc_client import HarvesterRpcClient
 from chia.rpc.rpc_server import start_rpc_server
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.bech32m import decode_puzzle_hash, encode_puzzle_hash
-from chia.util.block_tools import get_plot_dir
+from tests.block_tools import get_plot_dir
 from chia.util.config import load_config
 from chia.util.hash import std_hash
 from chia.util.ints import uint8, uint16, uint32, uint64
