+++ tests/simulation/test_simulation.py
@@ -1,7 +1,7 @@
 import pytest
 
 from chia.types.peer_info import PeerInfo
-from chia.util.block_tools import BlockTools
+from tests.block_tools import BlockTools
 from chia.util.ints import uint16
 from tests.core.node_height import node_height_at_least
 from tests.setup_nodes import self_hostname, setup_full_node, setup_full_system, test_constants
