+++ tests/util/benchmark_cost.py
@@ -9,7 +9,7 @@ from chia.types.blockchain_format.program import Program, INFINITE_COST
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.ints import uint32
-from chia.util.wallet_tools import WalletTool
+from tests.wallet_tools import WalletTool
 from chia.wallet.derive_keys import master_sk_to_wallet_sk
 from chia.wallet.puzzles.p2_delegated_puzzle import puzzle_for_pk
