+++ tests/util/key_tool.py
@@ -6,7 +6,7 @@ from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.coin_solution import CoinSolution
 from chia.util.condition_tools import conditions_by_opcode, conditions_for_solution, pkm_pairs_for_conditions_dict
 from tests.core.make_block_generator import GROUP_ORDER, int_to_public_key
-from chia.util.block_tools import test_constants
+from tests.block_tools import test_constants
 
 
 class KeyTool(dict):
